remove the flag and some tests that are obsolete as part of the deleting old CVE data model code that used this flag
